feat: PoC Device output via Buttplug.io
This is really the bare minimum in terms of closed loop
control: we take the motor speed from the nogasm and
we convert it into vibration or rotation commands.

The orgasm button (originally in place to record an orgasm
event) causes output to max out for 30 seconds: having the nogasm turn off the power once you're over the edge
seems mean.